separate currency selector component from header
