+++ src/components/Header/Header.jsx
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import { graphql } from '@apollo/client/react/hoc';
 
 import {
   Container,
@@ -16,98 +15,14 @@ import {
   GreyBox,
 } from './header.style';
 
-import Select from '../Select';
 import MiniCart from '../MiniCart';
+import CategoryTabs from '../Tabs';
+import CurrencySelector from '../CurrencySelector';
 
 import logo from '../../assets/images/logo.svg';
 import cart from '../../assets/images/empty_cart.svg';
-import { FETCH_CURRENCIES } from '../../graphql/queries';
 import { connect } from 'react-redux';
-import { changeCurrency } from '../../redux';
 import { Link } from 'react-router-dom';
-import CategoryTabs from '../Tabs';
-
-class WrappedSelect extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      currency: null,
-    };
-  }
-
-  componentDidUpdate = (_, prevState) => {
-    if (prevState && prevState.currency === null) {
-      this.setState({
-        currency: this.props.data.currencies[0],
-      });
-    }
-    if (prevState.currency !== this.state.currency) {
-      this.props.changeCurrency(this.state.currency);
-    }
-  };
-
-  setCurrency = currency => {
-    // transform the currency obj to better syntax
-    const newObj = {
-      label: currency.selectValue,
-      symbol: currency.displayValue,
-    };
-    this.setState({ currency: newObj });
-  };
-
-  render() {
-    const { loading, error, currencies } = this.props.data;
-    if (loading)
-      return (
-        <Container>
-          <Wrap>
-            <p>Loading...</p>
-          </Wrap>
-        </Container>
-      );
-    if (error)
-      return (
-        <Container>
-          <Wrap>
-            <p>Error:(</p>
-          </Wrap>
-        </Container>
-      );
-
-    const categoriesList = currencies.map(({ symbol, label }) => {
-      return { displayValue: symbol, selectValue: label };
-    });
-
-    return (
-      <Select
-        values={categoriesList}
-        setSelect={this.setCurrency}
-        updateParent
-        passObj
-        anchor
-        top="65px"
-        pad="16px 8px"
-        width="60px"
-      />
-    );
-  }
-}
-
-const mapDispatchToProps = dispatch => {
-  return {
-    changeCurrency: currency => dispatch(changeCurrency(currency)),
-  };
-};
-
-// connect to redux
-const withWrappedSelect = connect(null, mapDispatchToProps)(WrappedSelect);
-
-// fetch currencies list to pass to Select component
-const withCurrenciesQuery = graphql(FETCH_CURRENCIES);
-
-// Enhance component.
-const CurrencySelect = withCurrenciesQuery(withWrappedSelect);
 
 class Header extends Component {
   constructor(props) {
@@ -149,7 +64,7 @@ class Header extends Component {
               </Link>
             </Middle>
             <Right>
-              <CurrencySelect />
+              <CurrencySelector />
               <CartContainer onClick={this.toggleCart} ref={this.cartIconRef}>
                 <CartWrap>
                   <CartIcon src={cart} alt="" />
